# swagger_server/models/get_item_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class GetItemResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, item_id: str=None, name: str=None, type: str=None, price: float=None, expiry_date: date=None):  # noqa: E501
        """GetItemResponse - a model defined in Swagger

        :param item_id: The item_id of this GetItemResponse.  # noqa: E501
        :type item_id: str
        :param name: The name of this GetItemResponse.  # noqa: E501
        :type name: str
        :param type: The type of this GetItemResponse.  # noqa: E501
        :type type: str
        :param price: The price of this GetItemResponse.  # noqa: E501
        :type price: float
        :param expiry_date: The expiry_date of this GetItemResponse.  # noqa: E501
        :type expiry_date: date
        """
        self.swagger_types = {
            'item_id': str,
            'name': str,
            'type': str,
            'price': float,
            'expiry_date': date
        }

        self.attribute_map = {
            'item_id': 'item_id',
            'name': 'name',
            'type': 'type',
            'price': 'price',
            'expiry_date': 'expiry_date'
        }
        self._item_id = item_id
        self._name = name
        self._type = type
        self._price = price
        self._expiry_date = expiry_date

    @classmethod
    def from_dict(cls, dikt) -> 'GetItemResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GetItemResponse of this GetItemResponse.  # noqa: E501
        :rtype: GetItemResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def item_id(self) -> str:
        """Gets the item_id of this GetItemResponse.


        :return: The item_id of this GetItemResponse.
        :rtype: str
        """
        return self._item_id

    @item_id.setter
    def item_id(self, item_id: str):
        """Sets the item_id of this GetItemResponse.


        :param item_id: The item_id of this GetItemResponse.
        :type item_id: str
        """

        self._item_id = item_id

    @property
    def name(self) -> str:
        """Gets the name of this GetItemResponse.


        :return: The name of this GetItemResponse.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this GetItemResponse.


        :param name: The name of this GetItemResponse.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def type(self) -> str:
        """Gets the type of this GetItemResponse.


        :return: The type of this GetItemResponse.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this GetItemResponse.


        :param type: The type of this GetItemResponse.
        :type type: str
        """
        if type is None:
            raise ValueError("Invalid value for `type`, must not be `None`")  # noqa: E501

        self._type = type

    @property
    def price(self) -> float:
        """Gets the price of this GetItemResponse.


        :return: The price of this GetItemResponse.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this GetItemResponse.


        :param price: The price of this GetItemResponse.
        :type price: float
        """
        if price is None:
            raise ValueError("Invalid value for `price`, must not be `None`")  # noqa: E501

        self._price = price

    @property
    def expiry_date(self) -> date:
        """Gets the expiry_date of this GetItemResponse.


        :return: The expiry_date of this GetItemResponse.
        :rtype: date
        """
        return self._expiry_date

    @expiry_date.setter
    def expiry_date(self, expiry_date: date):
        """Sets the expiry_date of this GetItemResponse.


        :param expiry_date: The expiry_date of this GetItemResponse.
        :type expiry_date: date
        """
        if expiry_date is None:
            raise ValueError("Invalid value for `expiry_date`, must not be `None`")  # noqa: E501

        self._expiry_date = expiry_date